requirements.txt
python-dotenv==0.9.1
requests==2.25.0